# openapi/v1.yaml
openapi: "3.1.0"
info:
  title: Open LOS API
  description: Open source B2B Lending CRM — headless API
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /v1/deals/{dealId}/spread:
    parameters:
      - $ref: "#/components/parameters/DealId"
    post:
      operationId: createSpread
      summary: Ingest financial data and compute spread
      tags: [underwriting]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                entity_id:
                  type: string
                  format: uuid
                period:
                  type: string
                  description: "e.g. FY2024, H1-2025"
          application/json:
            schema:
              type: object
              required: [entity_id, period, line_items]
              properties:
                entity_id:
                  type: string
                  format: uuid
                period:
                  type: string
                line_items:
                  type: array
                  items:
                    type: object
                    required: [category, label, amount]
                    properties:
                      category:
                        type: string
                      label:
                        type: string
                      amount:
                        type: integer
                        description: Minor units (cents/pence)
      responses:
        "200":
          description: Spread computed
          content:
            application/json:
              schema:
                type: object
                required: [spread_id, ratios]
                properties:
                  spread_id:
                    type: string
                    format: uuid
                  entity_id:
                    type: string
                    format: uuid
                  period:
                    type: string
                  line_items:
                    type: array
                    items:
                      type: object
                  ratios:
                    type: object
                    properties:
                      current_ratio:
                        type: number
                      debt_to_equity:
                        type: number
                      dscr:
                        type: number
                      gross_margin:
                        type: number
                      net_margin:
                        type: number
                      leverage:
                        type: number
  /v1/deals/{dealId}/ratios:
    parameters:
      - $ref: "#/components/parameters/DealId"
    get:
      operationId: getRatios
      summary: Get computed financial ratios for a deal
      tags: [underwriting]
      responses:
        "200":
          description: Ratio summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  ratios:
                    type: array
                    items:
                      type: object
                      properties:
                        entity_id:
                          type: string
                          format: uuid
                        period:
                          type: string
                        current_ratio:
                          type: number
                        debt_to_equity:
                          type: number
                        dscr:
                          type: number
                        gross_margin:
                          type: number

  # ── Evaluate (Underwriting) ────────────────────────────────────────────
  /v1/deals/{dealId}/evaluate:
    parameters:
      - $ref: "#/components/parameters/DealId"
    post:
      operationId: evaluateDeal
      summary: Trigger underwriting evaluation for a deal
      description: |
        Triggers underwriting evaluation for the deal using the provided policy
        configuration. Returns a structured UnderwritingRun conforming to
        `schemas/underwriting-run.schema.json`.

        Current implementation is the deterministic thin-agent rules path.
        `mode=full` is accepted for forward compatibility and currently falls
        back to the same rules evaluation as `mode=rules_only`.
      tags: [underwriting]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Evaluation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ── Covenants ──────────────────────────────────────────────────────────
...
components:
  parameters:
    DealId:
      name: dealId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    EvaluateRequest:
      type: object
      properties:
        policy:
          type: object
          description: Lender policy configuration for the evaluation
          required: [policy_id, model]
          properties:
            policy_id:
              type: string
            model:
              type: string
              description: LLM model identifier
            persona:
              type: string
            target_yield_pct:
              type: number
            max_single_loan:
              type: number
            total_capital:
              type: number
            sector_limits:
              type: object
              additionalProperties:
                type: number
            params:
              type: object
              additionalProperties: true
        provider:
          type: string
          enum: [openrouter, anthropic]
          description: Reserved for future full-agent provider selection
        mode:
          type: string
          enum: [full, rules_only]
          default: full
          description: "full = LLM agent evaluation; rules_only = deterministic rules (thin agent fallback)"
        allow_rules_fallback:
          type: boolean
          default: false
          description: "Only when mode=full: if true, allows deterministic rules fallback when LLM is unavailable or fails"
    EvaluateResponse:
      type: object
      description: Structured UnderwritingRun result conforming to underwriting-run.schema.json
      required: [run_id, decision]
      properties:
        run_id:
          type: string
          format: uuid
        decision:
          type: object
          required: [action]
          properties:
            action:
              type: string
              enum: [approve, decline, counter, refer]
            terms:
              type: object
              properties:
                amount:
                  type: number
                apr:
                  type: number
                  description: "Decimal (0.095 = 9.5%)"
                tenor_months:
                  type: integer
            risk_grade:
              type: string
            confidence:
              type: number
        rationale:
          type: object
          properties:
            summary:
              type: string
            key_factors:
              type: array
              items:
                type: string
            what_would_change:
              type: array
              items:
                type: string
        trace:
          type: object
          properties:
            steps:
              type: array
              items:
                type: object
            latency_ms:
              type: integer
            cost:
              type: object
              properties:
                tokens_in:
                  type: integer
                tokens_out:
                  type: integer
                estimated_cost_usd:
                  type: number

    # ── Error ──
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            retryable:
              type: boolean
              default: false